requirements.txt
berserk-downstream==0.11.3
certifi==2020.12.5
chardet==4.0.0
chess==1.5.0
Deprecated==1.2.12
idna==2.10
ndjson==0.3.1
requests==2.25.1
urllib3==1.26.4
wrapt==1.12.1
